Switch from digital to analog read for screen photodiode
With the amplified photodiode for screen. The amplification was selected to
never reach saturation but this also means values below digital 1 for even the
brightest screens. Switched to watching the photodiode reading via analogRead.

We also need to add a way to measure and send out screen brightness as a
function of time during transition. Most screens take several milliseconds to
transition and we want to know the shape of this curve as seen by the
photodiode.